pyproject.toml
[tool.poetry]
name = "pathsql"
version = "0.1.0"
description = ""
authors = ["patarapolw <patarapolw@gmail.com>"]

[tool.poetry.dependencies]
python = "*"
sqlalchemy = "^1.2"
jupyter = "^1.0"
notebook = "^5.6"
pyexcel = "^0.5.9"
pyexcel-xls = "^0.5.8"
requests = "^2.19"
bs4 = "^0.0.1"
alembic = "^1.0"
psycopg2-binary = "^2.7"
sqlalchemy-citext = "^1.3"

[tool.poetry.dev-dependencies]
tinydb = "^3.11"
